+++ src/khmer/_cpy_graphlabels.cc
@@ -58,6 +58,7 @@ PyMethodDef khmer_graphlabels_methods[] = {
     {"consume_partitioned_fasta_and_tag_with_labels", (PyCFunction)labelhash_consume_partitioned_fasta_and_tag_with_labels, METH_VARARGS, "" },
     {"sweep_tag_neighborhood", (PyCFunction)labelhash_sweep_tag_neighborhood, METH_VARARGS, "" },
     {"get_tag_labels", (PyCFunction)labelhash_get_tag_labels, METH_VARARGS, ""},
+    {"link_tag_and_label", (PyCFunction)labelhash_link_tag_and_label, METH_VARARGS, ""},
     {"consume_sequence_and_tag_with_labels", (PyCFunction)labelhash_consume_sequence_and_tag_with_labels, METH_VARARGS, "" },
     {"n_labels", (PyCFunction)labelhash_n_labels, METH_VARARGS, ""},
     {"get_all_labels", (PyCFunction)labelhash_get_all_labels, METH_VARARGS, "" },
@@ -84,8 +85,8 @@ void khmer_graphlabels_dealloc(khmer_KGraphLabels_Object * obj)
     Py_TYPE(obj)->tp_free((PyObject*)obj);
 }
 
- PyObject * khmer_graphlabels_new(PyTypeObject *type, PyObject *args,
-                                        PyObject *kwds)
+PyObject * khmer_graphlabels_new(PyTypeObject *type, PyObject *args,
+                                 PyObject *kwds)
 {
     khmer_KGraphLabels_Object *self;
     self = (khmer_KGraphLabels_Object*)type->tp_alloc(type, 0);
@@ -94,6 +95,7 @@ void khmer_graphlabels_dealloc(khmer_KGraphLabels_Object * obj)
         PyObject * hashgraph_o;
         Hashgraph * hashgraph = NULL; // @CTB
 
+        // GraphLabels takes a single argument, a hashgraph descendant.
         if (!PyArg_ParseTuple(args, "O", &hashgraph_o)) {
             Py_DECREF(self);
             return NULL;
@@ -111,6 +113,10 @@ void khmer_graphlabels_dealloc(khmer_KGraphLabels_Object * obj)
             Py_DECREF(self);
             return NULL;
         }
+        // set 'base' for CPython-style inheritance.
+        self->khashgraph.khashtable.hashtable =
+            dynamic_cast<Hashtable*>(hashgraph);
+        self->khashgraph.hashgraph = dynamic_cast<Hashgraph*>(hashgraph);
 
         try {
             self->labelhash = new LabelHash(hashgraph);
@@ -387,6 +393,30 @@ labelhash_get_tag_labels(khmer_KGraphLabels_Object * me, PyObject * args)
     }
 
     return x;
+}
+
+
+PyObject *
+labelhash_link_tag_and_label(khmer_KGraphLabels_Object * me, PyObject * args)
+{
+    LabelHash * labelhash = me->labelhash;
+
+    PyObject * tag_o;
+    HashIntoType tag;
+    Label label;
+
+    if (!PyArg_ParseTuple(args, "OK", &tag_o, &label)) {
+        return NULL;
+    }
+    if (!ht_convert_PyObject_to_HashIntoType(tag_o, tag,
+            labelhash->graph)) {
+        return NULL;
+    }
+
+    labelhash->link_tag_and_label(tag, label);
+
+    Py_INCREF(Py_None);
+    return Py_None;
 }