+++ src/khmer/_cpy_khmer.cc
@@ -260,20 +260,6 @@ MOD_INIT(_khmer)
         return MOD_ERROR_VAL;
     }
 
-    khmer_KCounttable_Type.tp_base = &khmer_KHashtable_Type;
-    if (PyType_Ready(&khmer_KCounttable_Type) < 0) {
-        return MOD_ERROR_VAL;
-    }
-
-    khmer_KSmallCounttable_Type.tp_base = &khmer_KHashtable_Type;
-    if (PyType_Ready(&khmer_KSmallCounttable_Type) < 0) {
-        return MOD_ERROR_VAL;
-    }
-
-    khmer_KNodetable_Type.tp_base = &khmer_KHashtable_Type;
-    if (PyType_Ready(&khmer_KNodetable_Type) < 0) {
-        return MOD_ERROR_VAL;
-    }
 
     khmer_KHashgraph_Type.tp_base = &khmer_KHashtable_Type;
     khmer_KHashgraph_Type.tp_methods = khmer_hashgraph_methods;
@@ -313,9 +299,6 @@ MOD_INIT(_khmer)
         return MOD_ERROR_VAL;
     }
 
-    if (PyType_Ready(&khmer_KHLLCounter_Type) < 0) {
-        return MOD_ERROR_VAL;
-    }
     if (PyType_Ready(&khmer_ReadAlignerType) < 0) {
         return MOD_ERROR_VAL;
     }
@@ -366,23 +349,6 @@ MOD_INIT(_khmer)
         return MOD_ERROR_VAL;
     }
 
-    Py_INCREF(&khmer_KCounttable_Type);
-    if (PyModule_AddObject( m, "Counttable",
-                            (PyObject *)&khmer_KCounttable_Type ) < 0) {
-        return MOD_ERROR_VAL;
-    }
-
-    Py_INCREF(&khmer_KSmallCounttable_Type);
-    if (PyModule_AddObject( m, "SmallCounttable",
-                            (PyObject *)&khmer_KSmallCounttable_Type ) < 0) {
-        return MOD_ERROR_VAL;
-    }
-
-    Py_INCREF(&khmer_KNodetable_Type);
-    if (PyModule_AddObject( m, "Nodetable",
-                            (PyObject *)&khmer_KNodetable_Type ) < 0) {
-        return MOD_ERROR_VAL;
-    }
 
     Py_INCREF(&khmer_KCountgraph_Type);
     if (PyModule_AddObject( m, "Countgraph",
@@ -414,12 +380,6 @@ MOD_INIT(_khmer)
 
     khmer_HashSet_Type.tp_new = khmer_HashSet_new;
     if (PyType_Ready(&khmer_HashSet_Type) < 0) {
-        return MOD_ERROR_VAL;
-    }
-
-    Py_INCREF(&khmer_KHLLCounter_Type);
-    if (PyModule_AddObject(m, "HLLCounter",
-                           (PyObject *)&khmer_KHLLCounter_Type) < 0) {
         return MOD_ERROR_VAL;
     }