+++ src/oxli/storage.cc
@@ -914,3 +914,131 @@ void NibbleStorage::load(std::string infilename, WordLength& ksize)
         throw oxli_file_exception(err);
     }
 }
+
+
+void QFStorage::save(std::string outfilename, WordLength ksize)
+{
+    ofstream outfile(outfilename.c_str(), ios::binary);
+
+    unsigned char version = SAVED_FORMAT_VERSION;
+    unsigned char ht_type = SAVED_QFCOUNT;
+
+    outfile.write(SAVED_SIGNATURE, 4);
+    outfile.write((const char *) &version, 1);
+    outfile.write((const char *) &ht_type, 1);
+    outfile.write((const char *) &ksize, sizeof(ksize));
+
+    /* just a hack to handle __uint128_t value. Don't know a better to handle it
+     * right now */
+    uint64_t tmp_range;
+    tmp_range = cf.range;
+
+    outfile.write((const char *) &cf.nslots, sizeof(cf.nslots));
+    outfile.write((const char *) &cf.xnslots, sizeof(cf.xnslots));
+    outfile.write((const char *) &cf.key_bits, sizeof(cf.key_bits));
+    outfile.write((const char *) &cf.value_bits, sizeof(cf.value_bits));
+    outfile.write((const char *) &cf.key_remainder_bits, sizeof(cf.key_remainder_bits));
+    outfile.write((const char *) &cf.bits_per_slot, sizeof(cf.bits_per_slot));
+    outfile.write((const char *) &tmp_range, sizeof(tmp_range));
+    outfile.write((const char *) &cf.nblocks, sizeof(cf.nblocks));
+    outfile.write((const char *) &cf.nelts, sizeof(cf.nelts));
+    outfile.write((const char *) &cf.ndistinct_elts, sizeof(cf.ndistinct_elts));
+    outfile.write((const char *) &cf.noccupied_slots, sizeof(cf.noccupied_slots));
+
+    #if BITS_PER_SLOT == 8 || BITS_PER_SLOT == 16 || BITS_PER_SLOT == 32 || BITS_PER_SLOT == 64
+        outfile.write((const char *) cf.blocks, sizeof(qfblock) * cf.nblocks);
+    #else
+        outfile.write((const char *) cf.blocks,
+                      (sizeof(qfblock) + SLOTS_PER_BLOCK * cf.bits_per_slot / 8) * cf.nblocks);
+    #endif
+    outfile.close();
+}
+
+
+void QFStorage::load(std::string infilename, WordLength &ksize)
+{
+    ifstream infile;
+    // configure ifstream to raise exceptions for everything.
+    infile.exceptions(std::ifstream::failbit | std::ifstream::badbit |
+                      std::ifstream::eofbit);
+
+    try {
+        infile.open(infilename.c_str(), ios::binary);
+    } catch (std::ifstream::failure &e) {
+        std::string err;
+        if (!infile.is_open()) {
+            err = "Cannot open k-mer count file: " + infilename;
+        } else {
+            err = "Unknown error in opening file: " + infilename;
+        }
+        throw oxli_file_exception(err + " " + strerror(errno));
+    } catch (const std::exception &e) {
+        std::string err = "Unknown error opening file: " + infilename + " "
+                          + strerror(errno);
+        throw oxli_file_exception(err);
+    }
+    WordLength save_ksize = 0;
+    char signature [4];
+    unsigned char version = 0, ht_type = 0;
+    uint64_t tmp_range;
+
+    infile.read(signature, 4);
+    infile.read((char *) &version, 1);
+    infile.read((char *) &ht_type, 1);
+
+    if (!(std::string(signature, 4) == SAVED_SIGNATURE)) {
+        std::ostringstream err;
+        err << "Does not start with signature for a oxli file: 0x";
+        for(size_t i=0; i < 4; ++i) {
+            err << std::hex << (int) signature[i];
+        }
+        err << " Should be: " << SAVED_SIGNATURE;
+        throw oxli_file_exception(err.str());
+    } else if (!(version == SAVED_FORMAT_VERSION)) {
+        std::ostringstream err;
+        err << "Incorrect file format version " << (int) version
+            << " while reading k-mer count file from " << infilename
+            << "; should be " << (int) SAVED_FORMAT_VERSION;
+        throw oxli_file_exception(err.str());
+    } else if (!(ht_type == SAVED_QFCOUNT)) {
+        std::ostringstream err;
+        err << "Incorrect file format type " << (int) ht_type
+            << " expected " << (int) SAVED_QFCOUNT
+            << " while reading k-mer count file from " << infilename;
+        throw oxli_file_exception(err.str());
+    }
+
+    infile.read((char *) &save_ksize, sizeof(save_ksize));
+    ksize = save_ksize;
+
+    infile.read((char *) &cf.nslots, sizeof(cf.nslots));
+    infile.read((char *) &cf.xnslots, sizeof(cf.xnslots));
+    infile.read((char *) &cf.key_bits, sizeof(cf.key_bits));
+    infile.read((char *) &cf.value_bits, sizeof(cf.value_bits));
+    infile.read((char *) &cf.key_remainder_bits, sizeof(cf.key_remainder_bits));
+    infile.read((char *) &cf.bits_per_slot, sizeof(cf.bits_per_slot));
+    infile.read((char *) &tmp_range, sizeof(tmp_range));
+
+    infile.read((char *) &cf.nblocks, sizeof(cf.nblocks));
+    infile.read((char *) &cf.nelts, sizeof(cf.nelts));
+    infile.read((char *) &cf.ndistinct_elts, sizeof(cf.ndistinct_elts));
+    infile.read((char *) &cf.noccupied_slots, sizeof(cf.noccupied_slots));
+    /* just a hack to handle __uint128_t value. Don't know a better to handle it
+     * right now */
+    cf.range = tmp_range;
+    // deallocate previously allocated blocks
+    free(cf.blocks);
+    /* allocate the space for the actual qf blocks */
+    #if BITS_PER_SLOT == 8 || BITS_PER_SLOT == 16 || BITS_PER_SLOT == 32 || BITS_PER_SLOT == 64
+        cf.blocks = (qfblock *)calloc(cf.nblocks, sizeof(qfblock));
+    #else
+        cf.blocks = (qfblock *)calloc(cf.nblocks, sizeof(qfblock) + SLOTS_PER_BLOCK * cf.bits_per_slot / 8);
+    #endif
+    #if BITS_PER_SLOT == 8 || BITS_PER_SLOT == 16 || BITS_PER_SLOT == 32 || BITS_PER_SLOT == 64
+        infile.read((char *) cf.blocks, sizeof(qfblock) * cf.nblocks);
+    #else
+        infile.read((char *) cf.blocks,
+                    (sizeof(qfblock) + SLOTS_PER_BLOCK * cf.bits_per_slot / 8) * cf.nblocks);
+    #endif
+    infile.close();
+}