+++ CHANGELOG.md
@@ -12,9 +12,15 @@ under semantic versioning, but will be in future versions of khmer.
 - Cython wrapper for liboxli.
 - Cython containers for parsing, assembly, and hashing.
 - Header install for liboxli.
-- New function `consume_fasta_banding` for bulk loading of sequences into
-  hashtables. Only k-mers whose hashed values fall within a specified range are
-  counted.
+- New storage class using a Counting Quotient Filter with improved cache
+  locality over bloom filters.
+- New variants of the sequence bulk loading method with a "banding" mode and a
+  "mask" mode. In "banding" mode, only k-mers whose hashed values fall within a
+  specified range are counted. In "mask" mode, only k-mers not already pressent
+  in the specified mask are counted.
+    - `consume_seqfile_banding`
+    - `consume_seqfile_with_mask`
+    - `consume_seqfile_banding_with_mask`
 
 ### Changed
 - Non-ACTG handling significantly changed so that only bulk-loading functions